correct after first review

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -5,8 +5,8 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
 import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -26,27 +26,44 @@ public class InMemoryMealRepository implements MealRepository {
 
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(userId);
             repository.put(meal.getId(), meal);
             return meal;
+        } else if (meal.getUserId() == userId) {
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
         }
-        // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        return null;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return repository.remove(id) != null;
+        if (repository.get(id).getUserId() == userId) {
+            return repository.remove(id) != null;
+        }
+        return false;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        return repository.get(id);
+        if (repository.get(id).getUserId() == userId) {
+            return repository.get(id);
+        }
+        return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
+    public List<Meal> getAll(int userId) {
+        Comparator<Meal> comparator = new Comparator<Meal>() {
+            @Override
+            public int compare(Meal meal, Meal mealAnother) {
+                return meal.getDate().compareTo(mealAnother.getDate());
+            }
+        };
 
-        return repository.values().stream().sorted(Comparator.reverseOrder()).collect(Collectors.toList());
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == userId)
+                .sorted(comparator.reversed())
+                .collect(Collectors.toList());
     }
 }
